o Add unit test for merging that demonstrates bug updating references to   merged vertices when vertices are contained in tracking sets.
o Fix above bug

o Fix incorrect error handling in MBCore::merge_entities: if updating
  of references fails, don't delete the entity.



git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@3031 6162379e-bd28-0410-9a7d-b7f4dcfcad3c

+++ MBCore.cpp
@@ -1934,7 +1934,7 @@ MBErrorCode MBCore::merge_entities( MBEntityHandle entity_to_keep,
 
   result = aEntityFactory->merge_adjust_adjacencies(entity_to_keep, entity_to_remove);
   
-  if (delete_removed_entity) 
+  if (MB_SUCCESS == result && delete_removed_entity) 
     result = delete_entities(&entity_to_remove, 1);
 
   return result;
